% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwdata-package.R
\docType{package}
\name{cwdata-package}
\alias{cwdata}
\alias{cwdata-package}
\title{cwdata: Communicating with Data Resources}
\description{
Contains data and examples from CWD course.
}
\author{
\strong{Maintainer}: Zeyu Liu \email{zyliu16@student.monash.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
