% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_crop_yield.R
\docType{data}
\name{key_crop_yields}
\alias{key_crop_yields}
\title{Global key crop yield}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13075 rows and 15 columns.
}
\source{
The data below comes from Our World in Data and source from Tidy Tuesday.
}
\usage{
key_crop_yields
}
\description{
This data set contains crop yields from 249 countries
}
\details{
TODO: Add description of the columns
}
\keyword{datasets}
